Split `PurchasesTests` step 2: extracted `PurchasesConfiguringTests` and `PurchasesGetCustomerInfoTests`
Follow up to #1609.
