+++ Tests/UnitTests/Purchasing/Purchases/PurchasesTests.swift
@@ -18,108 +18,6 @@ import XCTest
 
 class PurchasesTests: BasePurchasesTests {
 
-    func testIsAbleToBeInitialized() {
-        setupPurchases()
-        expect(self.purchases).toNot(beNil())
-    }
-
-    func testUsingSharedInstanceWithoutInitializingThrowsAssertion() {
-        let expectedMessage = "Purchases has not been configured. Please call Purchases.configure()"
-        expectFatalError(expectedMessage: expectedMessage) { _ = Purchases.shared }
-    }
-
-    func testUsingSharedInstanceAfterInitializingDoesntThrowAssertion() {
-        setupPurchases()
-        expectNoFatalError { _ = Purchases.shared }
-    }
-
-    func testIsConfiguredReturnsCorrectvalue() {
-        expect(Purchases.isConfigured) == false
-        setupPurchases()
-        expect(Purchases.isConfigured) == true
-    }
-
-    func testConfigurationPassedThroughTimeouts() {
-        let networkTimeoutSeconds: TimeInterval = 9
-        let configurationBuilder = Configuration.Builder(withAPIKey: "")
-            .with(networkTimeout: networkTimeoutSeconds)
-            .with(storeKit1Timeout: networkTimeoutSeconds)
-        let purchases = Purchases.configure(with: configurationBuilder.build())
-
-        expect(purchases.networkTimeout) == networkTimeoutSeconds
-        expect(purchases.storeKitTimeout) == networkTimeoutSeconds
-    }
-
-    func testFirstInitializationCallDelegate() {
-        setupPurchases()
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(1))
-    }
-
-    func testFirstInitializationFromForegroundDelegateForAnonIfNothingCached() {
-        systemInfo.stubbedIsApplicationBackgrounded = false
-        setupPurchases()
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(1))
-    }
-
-    func testFirstInitializationFromBackgroundDoesntCallDelegateForAnonIfNothingCached() {
-        systemInfo.stubbedIsApplicationBackgrounded = true
-        setupPurchases()
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(0))
-    }
-
-    func testFirstInitializationFromBackgroundCallsDelegateForAnonIfInfoCached() throws {
-        systemInfo.stubbedIsApplicationBackgrounded = true
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object = try info.asData()
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-
-        setupPurchases()
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(1))
-    }
-
-    func testFirstInitializationFromBackgroundDoesntUpdateCustomerInfoCache() {
-        systemInfo.stubbedIsApplicationBackgrounded = true
-        setupPurchases()
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(0))
-    }
-
-    func testFirstInitializationFromForegroundUpdatesCustomerInfoCacheIfNotInUserDefaults() {
-        systemInfo.stubbedIsApplicationBackgrounded = false
-        setupPurchases()
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(1))
-    }
-
-    func testFirstInitializationFromForegroundUpdatesCustomerInfoCacheIfUserDefaultsCacheStale() {
-        let staleCacheDateForForeground = Calendar.current.date(byAdding: .minute, value: -20, to: Date())!
-        self.deviceCache.setCustomerInfoCache(timestamp: staleCacheDateForForeground,
-                                              appUserID: identityManager.currentAppUserID)
-        systemInfo.stubbedIsApplicationBackgrounded = false
-
-        setupPurchases()
-
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(1))
-    }
-
-    func testFirstInitializationFromForegroundUpdatesCustomerInfoEvenIfCacheValid() {
-        let staleCacheDateForForeground = Calendar.current.date(byAdding: .minute, value: -2, to: Date())!
-        self.deviceCache.setCustomerInfoCache(timestamp: staleCacheDateForForeground,
-                                              appUserID: identityManager.currentAppUserID)
-
-        systemInfo.stubbedIsApplicationBackgrounded = false
-
-        setupPurchases()
-
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(1))
-    }
-
     func testDelegateIsCalledForRandomPurchaseSuccess() throws {
         setupPurchases()
 
@@ -1299,173 +1197,6 @@ class PurchasesTests: BasePurchasesTests {
         expect(self.deviceCache.cacheCustomerInfoCount).toEventually(equal(2))
     }
 
-    func testCachedCustomerInfoHasSchemaVersion() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object = try info.asData()
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-        self.backend.timeout = true
-
-        setupPurchases()
-
-        var receivedInfo: CustomerInfo?
-
-        purchases.getCustomerInfo { (info, _) in
-            receivedInfo = info
-        }
-
-        expect(receivedInfo).toNot(beNil())
-        expect(receivedInfo?.schemaVersion).toNot(beNil())
-    }
-
-    func testCachedCustomerInfoHandlesNullSchema() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object = try info.asData(withNewSchemaVersion: NSNull())
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-        self.backend.timeout = true
-
-        setupPurchases()
-
-        var receivedInfo: CustomerInfo?
-
-        purchases.getCustomerInfo { (info, _) in
-            receivedInfo = info
-        }
-
-        expect(receivedInfo).to(beNil())
-    }
-
-    func testSendsCachedCustomerInfoToGetter() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object = try info.asData()
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-        self.backend.timeout = true
-
-        setupPurchases()
-
-        var receivedInfo: CustomerInfo?
-
-        purchases.getCustomerInfo { (info, _) in
-            receivedInfo = info
-        }
-
-        expect(receivedInfo).toNot(beNil())
-    }
-
-    func testCustomerInfoCompletionBlockCalledExactlyOnceWhenInfoCached() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object =  try info.asData()
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-        self.deviceCache.stubbedIsCustomerInfoCacheStale = true
-        self.backend.timeout = false
-
-        setupPurchases()
-
-        var callCount = 0
-
-        purchases.getCustomerInfo { (_, _) in
-            callCount += 1
-        }
-
-        expect(callCount).toEventually(equal(1))
-    }
-
-    func testDoesntSendsCachedCustomerInfoToGetterIfSchemaVersionDiffers() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object = try info.asData(withNewSchemaVersion: "bad_version")
-
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-        self.backend.timeout = true
-
-        setupPurchases()
-
-        var receivedInfo: CustomerInfo?
-
-        purchases!.getCustomerInfo { (info, _) in
-            receivedInfo = info
-        }
-
-        expect(receivedInfo).to(beNil())
-    }
-
-    func testDoesntSendsCachedCustomerInfoToGetterIfNoSchemaVersionInCached() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-
-        let object = try info.asData(withNewSchemaVersion: nil)
-
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-        self.backend.timeout = true
-
-        setupPurchases()
-
-        var receivedInfo: CustomerInfo?
-
-        purchases!.getCustomerInfo { (info, _) in
-            receivedInfo = info
-        }
-
-        expect(receivedInfo).to(beNil())
-    }
-
-    func testDoesntSendCacheIfNoCacheAndCallsBackendAgain() {
-        self.backend.timeout = true
-
-        setupPurchases()
-
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(1))
-
-        purchases!.getCustomerInfo { (_, _) in
-        }
-
-        expect(self.backend.getSubscriberCallCount).to(equal(2))
-    }
-
     func testFirstInitializationGetsOfferingsIfAppActive() {
         systemInfo.stubbedIsApplicationBackgrounded = false
         setupPurchases()
@@ -1538,44 +1269,6 @@ class PurchasesTests: BasePurchasesTests {
         expect(self.backend.invokedPostAttributionDataParametersList[0].data?.keys.contains("rc_idfv")) == true
         expect(self.backend.invokedPostAttributionDataParametersList[0].network) == AttributionNetwork.appleSearchAds
         expect(self.backend.invokedPostAttributionDataParametersList[0].appUserID) == self.purchases?.appUserID
-    }
-
-    func testSharedInstanceIsSetWhenConfiguring() {
-        let purchases = Purchases.configure(withAPIKey: "")
-        expect(Purchases.shared) === purchases
-    }
-
-    func testSharedInstanceIsSetWhenConfiguringWithConfiguration() {
-        let configurationBuilder = Configuration.Builder(withAPIKey: "")
-        let purchases = Purchases.configure(with: configurationBuilder.build())
-        expect(Purchases.shared) === purchases
-    }
-
-    func testSharedInstanceIsSetWhenConfiguringWithAppUserID() {
-        let purchases = Purchases.configure(withAPIKey: "", appUserID: "")
-        expect(Purchases.shared) === purchases
-    }
-
-    func testSharedInstanceIsSetWhenConfiguringWithObserverMode() {
-        let purchases = Purchases.configure(withAPIKey: "", appUserID: "", observerMode: true)
-        expect(Purchases.shared) === purchases
-        expect(Purchases.shared.finishTransactions) == false
-    }
-
-    func testSharedInstanceIsSetWhenConfiguringWithAppUserIDAndUserDefaults() {
-        let purchases = Purchases.configure(withAPIKey: "", appUserID: "", observerMode: false, userDefaults: nil)
-        expect(Purchases.shared) === purchases
-        expect(Purchases.shared.finishTransactions) == true
-    }
-
-    func testSharedInstanceIsSetWhenConfiguringWithAppUserIDAndUserDefaultsAndUseSK2() {
-        let purchases = Purchases.configure(withAPIKey: "",
-                                            appUserID: "",
-                                            observerMode: false,
-                                            userDefaults: nil,
-                                            useStoreKit2IfAvailable: true)
-        expect(Purchases.shared) === purchases
-        expect(Purchases.shared.finishTransactions) == true
     }
 
     func testWhenNoReceiptDataReceiptIsRefreshed() {
@@ -2117,17 +1810,6 @@ class PurchasesTests: BasePurchasesTests {
         self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
     }
 
-    func testFetchCustomerInfoWhenCacheStale() {
-        setupPurchases()
-        self.deviceCache.stubbedIsCustomerInfoCacheStale = true
-
-        self.purchases?.getCustomerInfo { (_, _) in
-
-        }
-
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(2))
-    }
-
     func testIsAnonymous() {
         setupAnonPurchases()
         expect(self.purchases.isAnonymous).to(beTrue())
@@ -2216,83 +1898,6 @@ class PurchasesTests: BasePurchasesTests {
 
         expect(self.backend.postReceiptDataCalled).to(beTrue())
         expect(self.backend.postedObserverMode).to(beFalse())
-    }
-
-    func testInvalidateCustomerInfoCacheRemovesCachedCustomerInfo() {
-        setupPurchases()
-        guard let nonOptionalPurchases = purchases else { fatalError("failed when setting up purchases for testing") }
-        let appUserID = identityManager.currentAppUserID
-        self.deviceCache.cache(customerInfo: Data(), appUserID: appUserID)
-        expect(self.deviceCache.cachedCustomerInfoData(appUserID: appUserID)).toNot(beNil())
-        expect(self.deviceCache.invokedClearCustomerInfoCacheCount) == 0
-
-        nonOptionalPurchases.invalidateCustomerInfoCache()
-        expect(self.deviceCache.cachedCustomerInfoData(appUserID: appUserID)).to(beNil())
-        expect(self.deviceCache.invokedClearCustomerInfoCacheCount) == 1
-    }
-
-    func testGetCustomerInfoAfterInvalidatingDoesntReturnCachedVersion() throws {
-        setupPurchases()
-        guard let nonOptionalPurchases = purchases else { fatalError("failed when setting up purchases for testing") }
-
-        let appUserID = identityManager.currentAppUserID
-        let oldAppUserInfo = Data()
-        self.deviceCache.cache(customerInfo: oldAppUserInfo, appUserID: appUserID)
-        let overrideCustomerInfo = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:]
-            ]])
-        self.backend.overrideCustomerInfoResult = .success(overrideCustomerInfo)
-
-        var receivedCustomerInfo: CustomerInfo?
-        var completionCallCount = 0
-        var receivedError: Error?
-        nonOptionalPurchases.getCustomerInfo { (customerInfo, error) in
-            completionCallCount += 1
-            receivedError = error
-            receivedCustomerInfo = customerInfo
-        }
-
-        nonOptionalPurchases.invalidateCustomerInfoCache()
-
-        expect(completionCallCount).toEventually(equal(1))
-        expect(receivedError).to(beNil())
-        expect(receivedCustomerInfo) == overrideCustomerInfo
-        expect(self.purchasesDelegate.customerInfoReceivedCount) == 1
-    }
-
-    func testGetCustomerInfoAfterInvalidatingCallsCompletionWithErrorIfBackendError() {
-        let backendError: BackendError = .networkError(
-            .unexpectedResponse(nil)
-        )
-        self.backend.overrideCustomerInfoResult = .failure(backendError)
-
-        setupPurchases()
-        expect(self.purchasesDelegate.customerInfoReceivedCount) == 0
-
-        let appUserID = identityManager.currentAppUserID
-        let oldAppUserInfo = Data()
-        self.deviceCache.cache(customerInfo: oldAppUserInfo, appUserID: appUserID)
-
-        var receivedCustomerInfo: CustomerInfo?
-        var completionCallCount = 0
-        var receivedError: Error?
-        purchases.getCustomerInfo { (customerInfo, error) in
-            completionCallCount += 1
-            receivedError = error
-            receivedCustomerInfo = customerInfo
-        }
-
-        purchases.invalidateCustomerInfoCache()
-
-        expect(completionCallCount).toEventually(equal(1))
-        expect(receivedError).toNot(beNil())
-        expect(receivedCustomerInfo).to(beNil())
-        expect(self.purchasesDelegate.customerInfoReceivedCount) == 0
     }
 
     func testInvalidateCustomerInfoCacheDoesntClearOfferingsCache() {
